thematic via spoofing jupyter output
i didn't find any way to emit markdown within a cell, only results: asis
so here is a brute force implementation of dark mode for thematic
the approach could be cleaned up for users

i'll need to rename this repository if this holds. but also note that
we could use a more robust structure e.g.

::: {.cell}
::: {.cell-output-display .quarto-light-marker}
![](thematic_files/figure-html/unnamed-chunk-2-1.png){width=672}
:::

::: {.cell-output-display .quarto-dark-marker}
![](thematic_files/figure-html/unnamed-chunk-2-2.png){width=672}
:::
:::